refactor(module): reorganize Excel package
