+++ modules/fbs-core/api/src/main/kotlin/de/thm/ii/fbs/model/v2/checker/excel/configuration/ExcelCheckerConfiguration.kt
@@ -1,4 +1,4 @@
-package de.thm.ii.fbs.model.v2.checker.excel
+package de.thm.ii.fbs.model.v2.checker.excel.configuration
 
 data class ExcelCheckerConfiguration(
     val tasks: List<ExcelTaskConfiguration>,
