+++ modules/fbs-core/api/src/main/kotlin/de/thm/ii/fbs/model/v2/checker/excel/configuration/ExcelTaskConfiguration.kt
@@ -1,0 +1,9 @@
+package de.thm.ii.fbs.model.v2.checker.excel.configuration
+
+data class ExcelTaskConfiguration(
+    val sheetIdx: Int,
+    val changeFields: List<ExcelTaskConfigurationChange> = emptyList(),
+    val checkFields: List<ExcelTaskConfigurationCheck> = emptyList(),
+    val name: String,
+    val hideInvalidFields: Boolean = false
+)
