+++ modules/fbs-core/api/src/main/kotlin/de/thm/ii/fbs/model/v2/checker/excel/configuration/ExcelTaskConfigurationChange.kt
@@ -1,0 +1,3 @@
+package de.thm.ii.fbs.model.v2.checker.excel.configuration
+
+data class ExcelTaskConfigurationChange(val cell: String, val newValue: String, val sheetIdx: Int)
