+++ modules/fbs-core/api/src/main/kotlin/de/thm/ii/fbs/model/v2/checker/excel/configuration/ExcelTaskConfigurationCheck.kt
@@ -1,0 +1,7 @@
+package de.thm.ii.fbs.model.v2.checker.excel.configuration
+
+data class ExcelTaskConfigurationCheck(
+    val range: String,
+    val hideInvalidFields: Boolean = false,
+    val errorMsg: String = ""
+)
