+++ modules/fbs-core/api/src/main/kotlin/de/thm/ii/fbs/model/v2/checker/excel/handler/ManualFeedbackHandler.kt
@@ -1,8 +1,8 @@
 package de.thm.ii.fbs.model.v2.checker.excel.handler
 
-import de.thm.ii.fbs.model.v2.checker.excel.AnalysisResult
-import de.thm.ii.fbs.model.v2.checker.excel.ExcelCheckerConfiguration
+import de.thm.ii.fbs.model.v2.checker.excel.configuration.ExcelCheckerConfiguration
 import de.thm.ii.fbs.model.v2.checker.excel.handler.context.ErrorAnalysisContext
+import de.thm.ii.fbs.model.v2.checker.excel.result.AnalysisResult
 import de.thm.ii.fbs.utils.v2.handler.Handle
 import de.thm.ii.fbs.utils.v2.handler.When
 import de.thm.ii.fbs.utils.v2.spreadsheet.SpreadsheetUtils.Companion.rangeToCells
