+++ modules/fbs-core/api/src/main/kotlin/de/thm/ii/fbs/model/v2/checker/excel/result/AnalysisResult.kt
@@ -1,4 +1,6 @@
-package de.thm.ii.fbs.model.v2.checker.excel
+package de.thm.ii.fbs.model.v2.checker.excel.result
+
+import de.thm.ii.fbs.model.v2.checker.excel.Cell
 
 data class AnalysisResult(
     private val errorCellResults: HashMap<Cell, CellResult> = HashMap()
@@ -16,7 +18,6 @@ data class AnalysisResult(
     fun getAllErrorCells(): Set<Cell> {
         return errorCellResults.keys
     }
-
 
     fun getErrorCells(): Set<Cell> {
         return errorCellResults.entries.filter { entry -> !entry.value.isPropagated }.map { entry -> entry.key }.toSet()
