+++ modules/fbs-core/api/src/main/kotlin/de/thm/ii/fbs/services/v2/checker/excel/ExcelCheckerServiceV2.kt
@@ -1,9 +1,13 @@
 package de.thm.ii.fbs.services.v2.checker.excel
 
-import de.thm.ii.fbs.model.v2.checker.excel.*
+import de.thm.ii.fbs.model.v2.checker.excel.Cell
+import de.thm.ii.fbs.model.v2.checker.excel.configuration.ExcelCheckerConfiguration
+import de.thm.ii.fbs.model.v2.checker.excel.graph.ErrorAnalysisSolution
 import de.thm.ii.fbs.model.v2.checker.excel.handler.ErrorHandler
 import de.thm.ii.fbs.model.v2.checker.excel.handler.ManualFeedbackHandler
 import de.thm.ii.fbs.model.v2.checker.excel.handler.PropagatedErrorHandler
+import de.thm.ii.fbs.model.v2.checker.excel.result.AnalysisResult
+import de.thm.ii.fbs.model.v2.checker.excel.result.ExcelCheckerResultData
 import de.thm.ii.fbs.services.v2.handler.HandlerService
 import org.apache.poi.xssf.usermodel.XSSFWorkbook
 import org.springframework.stereotype.Service
