+++ modules/fbs-core/api/src/main/kotlin/de/thm/ii/fbs/services/v2/checker/storage/CheckerStorageService.kt
@@ -2,7 +2,7 @@ package de.thm.ii.fbs.services.v2.checker.storage
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.module.SimpleModule
-import de.thm.ii.fbs.model.v2.checker.excel.ReferenceGraph
+import de.thm.ii.fbs.model.v2.checker.excel.graph.ReferenceGraph
 import de.thm.ii.fbs.model.v2.checker.storage.CheckerStorageEntity
 import de.thm.ii.fbs.services.v2.persistence.CheckerStorageRepository
 import de.thm.ii.fbs.utils.v2.converters.ReferenceGraphDeserializer
