+++ modules/fbs-core/api/src/main/kotlin/de/thm/ii/fbs/utils/v2/converters/ReferenceGraphDeserializer.kt
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.core.JsonParser
 import com.fasterxml.jackson.databind.DeserializationContext
 import com.fasterxml.jackson.databind.JsonDeserializer
 import com.fasterxml.jackson.databind.JsonNode
-import de.thm.ii.fbs.model.v2.checker.excel.ReferenceGraph
+import de.thm.ii.fbs.model.v2.checker.excel.graph.ReferenceGraph
 
 class ReferenceGraphDeserializer : JsonDeserializer<ReferenceGraph>() {
     override fun deserialize(jp: JsonParser, ctxt: DeserializationContext?): ReferenceGraph {
