+++ modules/fbs-core/api/src/main/kotlin/de/thm/ii/fbs/utils/v2/converters/ReferenceGraphSerializer.kt
@@ -3,7 +3,7 @@ package de.thm.ii.fbs.utils.v2.converters
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.databind.JsonSerializer
 import com.fasterxml.jackson.databind.SerializerProvider
-import de.thm.ii.fbs.model.v2.checker.excel.ReferenceGraph
+import de.thm.ii.fbs.model.v2.checker.excel.graph.ReferenceGraph
 
 class ReferenceGraphSerializer : JsonSerializer<ReferenceGraph>() {
     override fun serialize(
