+++ modules/fbs-core/api/src/main/scala/de/thm/ii/fbs/services/checker/excel/ExcelCheckerService.scala
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.databind.{DeserializationFeature, ObjectMapper}
 import de.thm.ii.fbs.model._
 import de.thm.ii.fbs.model.checker.excel.SpreadsheetCell
 import de.thm.ii.fbs.model.task.Task
-import de.thm.ii.fbs.model.v2.checker.excel.ExcelCheckerResultData
+import de.thm.ii.fbs.model.v2.checker.excel.result.ExcelCheckerResultData
 import de.thm.ii.fbs.services.checker.`trait`.{CheckerService, CheckerServiceOnMainFileUpload, CheckerServiceOnSecondaryFileUpload}
 import de.thm.ii.fbs.services.persistence.{CheckrunnerSubTaskService, SubmissionService}
 import de.thm.ii.fbs.services.v2.checker.excel.{ErrorAnalysisSolutionService, ExcelCheckerServiceV2}
